# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /earn/adsgram/reward:
    post:
      operationId: claimAdsgramReward
      tags: [earn]
      summary: Credit points for a completed rewarded ad view (server-trusted cooldown + daily cap)
      responses:
        "200":
          description: Reward credited
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EarnRewardResult"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "429":
          description: Cooldown active or daily cap reached
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /earn/offerwall/postback:
    get:
      operationId: offerwallPostback
      tags: [earn]
      summary: Server-to-server postback endpoint called by CPA/offerwall/survey providers to credit a completed offer
      parameters:
        - name: provider
          in: query
          required: true
          schema:
            type: string
        - name: telegramId
          in: query
          required: true
          schema:
            type: string
        - name: amount
          in: query
          required: true
          schema:
            type: number
        - name: secret
          in: query
          required: true
          schema:
            type: string
        - name: txId
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: Credited
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EarnRewardResult"
        "400":
          description: Invalid provider or missing params
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Invalid postback secret
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /stars/invoice:
    post:
      operationId: createStarsInvoice
      tags: [stars]
      summary: Create a Telegram Stars invoice link for a purchasable product
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StarsInvoiceInput"
      responses:
        "200":
          description: Invoice link created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StarsInvoiceResult"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "400":
          description: Unknown product or Stars purchases not configured
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    EarnRewardResult:
      type: object
      properties:
        creditedPoints:
          type: integer
        lifetimePoints:
          type: integer
      required:
        - creditedPoints
        - lifetimePoints
    StarsInvoiceInput:
      type: object
      properties:
        product:
          type: string
          enum: [energy_refill, boost, premium_month]
      required:
        - product
    StarsInvoiceResult:
      type: object
      properties:
        invoiceUrl:
          type: string
        priceStars:
          type: integer
      required:
        - invoiceUrl
        - priceStars
...