# lib/api-zod/src/generated/types/starsInvoiceInput.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { StarsInvoiceInputProduct } from './starsInvoiceInputProduct';

export interface StarsInvoiceInput {
  product: StarsInvoiceInputProduct;
}